refactor and style changes
refactor getCharacteristicList from handlebar to function
refactor getDamageTypes to sheet utilities
fix style issues with larger char tables for animals and robots
