+++ src/module/handlebars.ts
@@ -7,7 +7,7 @@ import { TWODSIX } from "./config";
 import TwodsixItem from "./entities/TwodsixItem";
 import {Skills} from "../types/template";
 import TwodsixActor, { getPower, getWeight } from "./entities/TwodsixActor";
-import { _getTranslatedCharacteristicList, _genUntranslatedCharacteristicList } from "./utils/TwodsixRollSettings";
+import { getCharacteristicList } from "./utils/TwodsixRollSettings";
 import { simplifySkillName } from "./utils/utils";
 
 export default function registerHandlebarsHelpers(): void {
@@ -409,13 +409,7 @@ export default function registerHandlebarsHelpers(): void {
   });
 
   Handlebars.registerHelper('getCharacteristicList', (actor: TwodsixActor) => {
-    let returnValue = {};
-    if (actor) {
-      returnValue = _getTranslatedCharacteristicList(actor);
-    } else {
-      returnValue = _genUntranslatedCharacteristicList();
-    }
-    return returnValue;
+    return getCharacteristicList(actor);
   });
 
   Handlebars.registerHelper('makePieImage', (text: string) => {
