+++ static/styles/twodsix.css
@@ -994,7 +994,7 @@ h2 {
 }
 
 .bgi-species input {
-  width: 8em;
+  width: 7em;
   border: none;
 }
